gh-65 Move all terminology related resources
