+++ src/index.ts
@@ -22,9 +22,8 @@ export * from './mdm-model-types.model';
 export * from './mdm-model-types.resource';
 export * from './mdm-container-types.model';
 export * from './data-models/index';
+export * from './terminologies/index';
 export * from './mdm-merge-types.model';
-export * from './mdm-term.model';
-export * from './mdm-term.resource';
 export * from './mdm-classifier.model';
 export * from './mdm-classifier.resource';
 export * from './mdm-resources-configuration';
@@ -37,8 +36,6 @@ export * from './mdm-provider.model';
 export * from './mdm-provider.resource';
 export * from './mdm-catalogue-item.model';
 export * from './mdm-catalogue-item.resource';
-export * from './mdm-code-set.model';
-export * from './mdm-code-set.resource';
 export * from './mdm-reference-data-model.model';
 export * from './mdm-reference-data-model.resource';
 export * from './mdm-reference-data-element.resource';
@@ -50,10 +47,6 @@ export * from './mdm-admin.resource';
 export * from './mdm-rest-handler';
 export * from './mdm-security.model';
 export * from './mdm-security.resource';
-export * from './mdm-terminology.model';
-export * from './mdm-terminology.resource';
-export * from './mdm-term-relationship-type.model';
-export * from './mdm-term-replationship-type.resource';
 export * from './mdm-data-class-component.resource';
 export * from './mdm-namespaces.resource';
 export * from './mdm-catalogue-user.resource';
