+++ src/mdm-merge.model.ts
@@ -20,7 +20,7 @@ SPDX-License-Identifier: Apache-2.0
 import { DataModelDetail } from './data-models/mdm-data-model.model';
 import { Branchable, Modelable } from './mdm-model-types.model';
 import { ReferenceDataModelDetail } from './mdm-reference-data-model.model';
-import { TerminologyDetail } from './mdm-terminology.model';
+import { TerminologyDetail } from './terminologies/mdm-terminology.model';
 import { VersionedFolderDetail } from './mdm-versioned-folder.model';
 import {
   MdmResponse,
