+++ src/terminologies/index.ts
@@ -1,0 +1,8 @@
+export * from './mdm-code-set.model';
+export * from './mdm-code-set.resource';
+export * from './mdm-terminology.model';
+export * from './mdm-terminology.resource';
+export * from './mdm-term.model';
+export * from './mdm-term.resource';
+export * from './mdm-term-relationship-type.model';
+export * from './mdm-term-relationship-type.resource';
