+++ src/terminologies/mdm-term-relationship-type.model.ts
@@ -17,8 +17,14 @@ limitations under the License.
 SPDX-License-Identifier: Apache-2.0
 */
 
-import { Historical, Securable } from './mdm-model-types.model';
-import { CatalogueItem, MdmIndexResponse, MdmResponse, Navigatable, Uuid } from './mdm-common.model';
+import { Historical, Securable } from '../mdm-model-types.model';
+import {
+  CatalogueItem,
+  MdmIndexResponse,
+  MdmResponse,
+  Navigatable,
+  Uuid
+} from '../mdm-common.model';
 import { TermDetail } from './mdm-term.model';
 
 export interface TermRelationshipType extends CatalogueItem, Navigatable {
@@ -30,13 +36,16 @@ export interface TermRelationshipType extends CatalogueItem, Navigatable {
   childRelationship: boolean;
 }
 
-export type TermRelationshipTypeDetail =
-  TermRelationshipType
-  & Securable
-  & Historical;
+export type TermRelationshipTypeDetail = TermRelationshipType &
+  Securable &
+  Historical;
 
-export type TermRelationshipTypeIndexResponse = MdmIndexResponse<TermRelationshipType>;
-export type TermRelationshipTypeDetailResponse = MdmResponse<TermRelationshipTypeDetail>;
+export type TermRelationshipTypeIndexResponse = MdmIndexResponse<
+  TermRelationshipType
+>;
+export type TermRelationshipTypeDetailResponse = MdmResponse<
+  TermRelationshipTypeDetail
+>;
 
 export interface TermRelationship extends CatalogueItem, Navigatable {
   [key: string]: any;
