+++ src/terminologies/mdm-terminology.model.ts
@@ -16,28 +16,17 @@ limitations under the License.
 
 SPDX-License-Identifier: Apache-2.0
 */
-import { Historical, Securable } from './mdm-model-types.model';
-import { CatalogueItemDomainType, MdmIndexResponse, MdmResponse, Navigatable, Uuid } from './mdm-common.model';
+import { Modelable, ModelDomainDetail } from '../mdm-model-types.model';
+import { MdmIndexResponse, MdmResponse } from '../mdm-common.model';
 
-export interface Term {
+export type TerminologyDataType = 'Terminology';
+
+export interface Terminology extends Modelable {
   [key: string]: any;
-  id?: Uuid;
-  domainType: CatalogueItemDomainType;
-  model?: Uuid;
-  code: string;
-  definition: string;
-
-  /**
-   * The combination of {@link Term.code} and {@link Term.definition}.
-   */
-  label?: string;
+  type?: TerminologyDataType;
 }
 
-export type TermDetail =
-  Term
-  & Navigatable
-  & Securable
-  & Historical;
+export type TerminologyDetail = Terminology & ModelDomainDetail;
 
-export type TermIndexResponse = MdmIndexResponse<Term>;
-export type TermDetailResponse = MdmResponse<TermDetail>;
+export type TerminologyIndexResponse = MdmIndexResponse<Terminology>;
+export type TerminologyDetailResponse = MdmResponse<TerminologyDetail>;
